Rename lib2 to lib_sdk
