+++ api_tests/lib_sdk/wallets/bitcoin.ts
