+++ api_tests/lib_sdk/wallets/index.ts
